Add CallNativeFunction - see description below
As the name suggests, it allows calling an arbitrary native function,
provided this function was defined in the python filterscript. This way,
a user can flexibly add new functions to the Python plugin without
editing or compiling any C code.
An example use-case (the one this feature was developed for) could be
allowing the use of streamer plugins.